% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{ataccc_logit_hazard_covar}
\alias{ataccc_logit_hazard_covar}
\title{A vector of the covariance of logit(hazard), starting from time 1}
\usage{
ataccc_logit_hazard_covar(estimate = "hakki")
}
\description{
A vector of the covariance of logit(hazard), starting from time 1
}
