% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{ataccc_logit_hazard_mean}
\alias{ataccc_logit_hazard_mean}
\title{A vector of the mean of logit(hazard), starting from time 1}
\usage{
ataccc_logit_hazard_mean(estimate = "hakki")
}
\description{
A vector of the mean of logit(hazard), starting from time 1
}
