% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{ataccc_posterior_means}
\alias{ataccc_posterior_means}
\title{The posterior mean of the survival time for each posterior sample}
\usage{
ataccc_posterior_means(estimate = "hakki")
}
\arguments{
\item{estimate}{The estimate to use, either "hakki" or "original"}
}
\description{
The posterior mean of the survival time for each posterior sample
}
