% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ataccc_posterior_samples}
\alias{ataccc_posterior_samples}
\title{Get the posterior samples of the duration estimates}
\usage{
ataccc_posterior_samples(estimate = "hakki")
}
\arguments{
\item{estimate}{The estimate to use, either "hakki" or "original"}
}
\description{
Get the posterior samples of the duration estimates
}
