% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{ataccc_posterior_summary_stats}
\alias{ataccc_posterior_summary_stats}
\title{Summary statistics for each day of the duration estimate}
\usage{
ataccc_posterior_summary_stats(
  estimate = "hakki",
  summary_fun = ggdist::mean_qi,
  ...
)
}
\arguments{
\item{estimate}{The estimate to use, either "hakki" or "original"}

\item{summary_fun}{The function to use to summarise the posterior samples}

\item{...}{Additional arguments to pass to the summary function}
}
\description{
Summary statistics for each day of the duration estimate
}
