% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacccDurationEstimates-package.R
\docType{package}
\name{atacccDurationEstimates-package}
\alias{atacccDurationEstimates}
\alias{atacccDurationEstimates-package}
\title{atacccDurationEstimates: Estimes of the Distribution of the Duration of RT-PCR Positivity based on ATACCCC data}
\description{
RT-PCR duration has been estimated by using a viral load model. This package contains the posterior estimates.
}
\author{
\strong{Maintainer}: Joshua Blake \email{joshua.blake@mrc-bsu.cam.ac.uk} (\href{https://orcid.org/0000-0001-7522-3741}{ORCID})

}
\keyword{internal}
